+++ src/Bladeburner/ui/Console.tsx
@@ -26,8 +26,18 @@ export function Console(props: IProps): React.ReactElement {
 
     // TODO: Figure out how to actually make the scrolling work correctly.
     function scrollToBottom(): void {
-        if(lastRef.current)
+        function isMaxed(): boolean {
+            if(!lastRef.current) return false;
+            const oldTop = lastRef.current.scrollTop;
             lastRef.current.scrollTop = lastRef.current.scrollHeight;
+            const maxed = oldTop === lastRef.current.scrollTop;
+            lastRef.current.scrollTop = oldTop;
+            return maxed;
+        }
+        if(lastRef.current) {
+            if(isMaxed())
+                lastRef.current.scrollTop = lastRef.current.scrollHeight;
+        }
     }
 
     function rerender(): void {